python/django-react/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
geojson==3.1.0
psycopg==3.1.15
psycopg2-binary==2.9.9
pyproj==3.6.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
